Spielsword class addition
-Added Dictionary of Villagers for spawning.
-Added prefab for each class for easier spawning of Villagers.
-Changed xDir into int for simplicitys sake.
-Better renaming of variables

+++ Assets/Scripts/Character.cs
@@ -17,7 +17,7 @@ public class Character : MonoBehaviour {
     /// None =  0
     /// Right=  1
     /// </summary>
-    public float xDir;
+    public int xDir;
 
     public bool Alive
     {
@@ -56,7 +56,7 @@ public class Character : MonoBehaviour {
     {
         animData = new Hashtable
         {
-            { "Move", 0f },
+            { "Move", 0 },
             { "Dead", false },
             { "MeleeAttack", false },
             { "RangedAttack", false },
